Fix: Cast processing_time and tokens_used to integers
diffInSeconds() can return float values which cause MySQL 'out of range'
errors when inserting into unsigned integer columns.

+++ app/Jobs/Analysis/BaseAnalysisJob.php
@@ -102,12 +102,12 @@ abstract class BaseAnalysisJob implements ShouldQueue
 
         $result = $this->getAIService()->complete($prompt, $options);
 
-        $processingTime = now()->diffInSeconds($startTime);
+        $processingTime = (int) now()->diffInSeconds($startTime);
 
         return [
             'result' => $result['content'],
             'processing_time' => $processingTime,
-            'tokens_used' => $result['usage']['total_tokens'] ?? 0,
+            'tokens_used' => (int) ($result['usage']['total_tokens'] ?? 0),
             'model' => $result['model'],
             'provider' => $result['provider'],
         ];
